+++ src/main/java/ru/mcmodding/tutorial/client/ClientProxy.java
@@ -5,6 +5,7 @@ import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.world.World;
 import ru.mcmodding.tutorial.client.gui.ChestGui;
+import ru.mcmodding.tutorial.client.handler.KeyHandler;
 import ru.mcmodding.tutorial.common.CommonProxy;
 import ru.mcmodding.tutorial.common.container.ChestContainer;
 import ru.mcmodding.tutorial.common.handler.ModBlocks;
@@ -19,6 +20,9 @@ public class ClientProxy extends CommonProxy {
 
         ModItems.registerRender();
         ModBlocks.registerRender();
+
+        KeyHandler keyHandler = new KeyHandler();
+        keyHandler.register();
     }
 
     @Override
